# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API GameDashboard",
    "version": "1.0.0",
    "description": "Panel de gestion de serveurs de jeu. Les erreurs suivent Problem Details (RFC 9457) : un corps JSON portant `title`, `status` et `detail`, et non un simple texte."
  },
  "servers": [
    {
      "url": "https://panel.example",
      "description": "Le panel, sur son domaine"
    }
  ],
...
  "paths": {
...
    "/api/v1/application/users/{user}": {
      "get": {
        "summary": "Détail d'un compte et nombre de serveurs possédés.",
        "description": "Clé de plateforme. Portée requise : `users.read`.",
        "tags": [
          "Comptes"
        ],
        "operationId": "application-get-users-by-user",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "description": "Identifiant du compte.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "applicationKey": [
              "users.read"
            ]
          }
        ],
        "x-gd-scope": "users.read",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « users.read »."
          },
          "2XX": {
            "description": "Résultat de la lecture."
          }
        }
      },
      "patch": {
        "summary": "Corriger nom, prénom ou identifiant externe. Ni rôle, ni adresse, ni mot de passe.",
        "description": "Clé de plateforme. Portée requise : `users.write`.",
        "tags": [
          "Comptes"
        ],
        "operationId": "application-patch-users-by-user",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "description": "Identifiant du compte.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "applicationKey": [
              "users.write"
            ]
          }
        ],
        "x-gd-scope": "users.write",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « users.write »."
          },
          "2XX": {
            "description": "Résultat de la lecture."
          }
        }
      },
      "delete": {
        "summary": "Supprimer un compte. Refusé tant qu'il possède des serveurs.",
        "description": "Clé de plateforme. Portée requise : `users.delete`.",
        "tags": [
          "Comptes"
        ],
        "operationId": "application-delete-users-by-user",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "description": "Identifiant du compte.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "applicationKey": [
              "users.delete"
            ]
          }
        ],
        "x-gd-scope": "users.delete",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « users.delete »."
          },
          "2XX": {
            "description": "Résultat de la lecture."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-gd_session",
        "description": "Cookie de session du panel. Posé à la connexion, illisible par un script, et donc inutilisable depuis un client tiers."
      },
      "clientKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Clé personnelle, créée depuis « Compte → Clés d'API ». Elle agit au nom de son propriétaire et ne porte jamais plus que les portées qu'on lui a données."
      },
      "applicationKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Clé de plateforme, créée depuis l'administration. Elle n'agit au nom de personne : elle provisionne, suspend et résilie pour le compte d'un système tiers."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "Refus ou erreur, au format Problem Details (RFC 9457).",
        "properties": {
          "title": {
            "type": "string",
            "description": "Ce qui s'est passé, en une phrase."
          },
          "status": {
            "type": "integer",
            "description": "Le code HTTP, répété dans le corps."
          },
          "detail": {
            "type": "string",
            "description": "Le détail utile, quand il y en a un."
          }
        },
        "required": [
          "title",
          "status"
        ]
      }
    }
  }
}